update(multilang): move folder to [locale]
