+++ src/app/[locale]/aboutus/page.tsx
@@ -56,3 +56,8 @@ export default function AboutusPage() {
     </main>
   );
 }
+
+const locales = ['en', 'th'];
+export function generateStaticParams() {
+  return locales.map((locale) => ({locale}));
+}
