+++ src/app/[locale]/privacy-policy/page.tsx
@@ -56,3 +56,8 @@ export default function PrivacyPolicyPage() {
     </main>
   );
 }
+
+const locales = ['en', 'th'];
+export function generateStaticParams() {
+  return locales.map((locale) => ({locale}));
+}
