+++ src/app/[locale]/projects/page.tsx
