+++ src/app/[locale]/services/page.tsx
@@ -4,9 +4,9 @@ import { PrimaryButton } from "@/components/buttons/PrimaryButton";
 import FormPage from "@/features/ContactPage/FormPage";
 import { Metadata } from "next";
 import { Poppins } from "next/font/google";
-import ServiceIntro from "../../features/ServicePage/components/ServiceIntro/ServiceIntro";
+import ServiceIntro from "../../../features/ServicePage/components/ServiceIntro/ServiceIntro";
 import ContactHomeSection from "@/components/ContactHomeSection/ContactHomeSection";
-import ServiceCard from "../../features/ServicePage/components/ServiceCard/ServiceCard";
+import ServiceCard from "../../../features/ServicePage/components/ServiceCard/ServiceCard";
 import SocialContactFloating from "@/components/SocialContactFloating/SocialContactFloating";
 import ServiceWrapper from "@/features/ServicePage/ServiceWrapper";
 
@@ -51,3 +51,9 @@ export default function ServicePage() {
     </main>
   );
 }
+
+const locales = ['en', 'th'];
+export function generateStaticParams() {
+  return locales.map((locale) => ({locale}));
+} 
+
